refactor: Facade 이름 Refactoring
